requirements.txt
django==5.2.9
djangorestframework>=3.14,<4.0
psycopg2-binary>=2.9,<3.0